Move `JsonResponse` to `Engine\Response` namespace and update all references
